- fix bug: only delete app cache if the file exists

+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/download/AppCache.kt
@@ -26,7 +26,9 @@ class AppCache(val app: App) {
 
     fun delete(context: Context) {
         val file = getFile(context)
-        val success = file.delete()
-        check(success) { "Fail to delete file '${file.absolutePath}'." }
+        if (file.exists()) {
+            val success = file.delete()
+            check(success) { "Fail to delete file '${file.absolutePath}'." }
+        }
     }
 }
